EType double

+++ tasks/konstantinov_s_elem_vec_sign_change_count/common/include/common.hpp
@@ -9,7 +9,7 @@
 
 namespace konstantinov_s_elem_vec_sign_change_count {
 
-using EType = int16_t;
+using EType = double;
 using InType = std::vector<EType>;
 using OutType = int;
 using TestType = int;  // std::tuple<std::tuple<std::vector<int>, int>, std::string>;
